requirements.txt
numpy==1.19.5
pandas==1.1.5
opencv-python==4.5.5.64
torchvision==0.11.2
tqdm==4.64.0
PyYAML==5.4.1
scikit-learn==0.24.2
scipy==1.5.4
timm==0.5.4
pytorch_metric_learning==1.4.0
tensorboard==2.9.1